third_party/blink/web_tests/accessibility/canvas-fallback-content-labels-expected.txt

This tests that labels and elements with labels in canvas fallback content have the same accessibility properties as those same elements outside of a canvas.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Checking accessibility properties for ids: nothing1, nothing2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: text_label1, text_label2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: text1, text2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: check_label1, check_label2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: check1, check2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: radio_label1, radio_label2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: radio1, radio2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: slider_label1, slider_label2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: slider1, slider2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: list_label1, list_label2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: list1, list2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: label_for1, label_for2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


Checking accessibility properties for ids: text_for1, text_for2
PASS axElement2.role is axElement1.role
PASS collapseWhitespace(axElement2.name) is collapseWhitespace(axElement1.name)
PASS axElement2.nameElementCount() is axElement1.nameElementCount()


PASS successfullyParsed is true

TEST COMPLETE
